% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_MCMC.R
\name{compare_MCMC}
\alias{compare_MCMC}
\title{Benchmark results}
\usage{
compare_MCMC(
  rj.dat = NULL,
  gvs.fixed = NULL,
  gvs.random = NULL,
  by.model = FALSE,
  kernel.adj = 2,
  viridis.col = FALSE,
  density = TRUE,
  prob = TRUE
)
}
\arguments{
\item{rj.dat}{Input rjMCMC object, as returned by \code{\link{trace_rjMCMC}}.}

\item{gvs.fixed}{Input Gibbs object, as returned by \code{\link{gibbs}} using a fixed effect implementation of the dose-response model.}

\item{gvs.random}{Input Gibbs object, as returned by \code{\link{gibbs}} using a random effect implementation of the dose-response model.}

\item{by.model}{Logical. If \code{TRUE}, the functions subsets posterior estimates by candidate model.}

\item{kernel.adj}{Bandwidth adjustment. The bandwidth used to create density plots is given by \code{kernel.adj*bw}.}

\item{viridis.col}{Logical. Whether to use a viridis colour scheme.}

\item{density}{Logical. If \code{TRUE}, compares density plots for each model parameter.}

\item{prob}{Logical. If \code{TRUE}, compares posterior rankings for candidate models.}
}
\description{
Evaluate outputs from models implemented using a (1) rjMCMC and (2) Gibbs Variable Selection approach. Returns comparative plots of posterior model rankings and posterior parameter estimates.
}
\examples{
library(espresso)

# Import the example data

mydat <- read_data(file = NULL) 

# Import a real dataset with the sonar and range covariates, 
# excluding sperm whales and any other species with a sample size
# smaller than two

mydat <- read_data(file = "path/to/my/data.csv", 
                  exclude.species = "Sperm whale",
                  min.N = 2) 

}
\seealso{
\code{\link{simulate_data}} \code{\link{example_brs}} \code{\link{summary.rjdata}}
}
\author{
Phil J. Bouchet
}
\keyword{brs}
\keyword{rjmcmc}
