% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_phase.R
\name{compare_phase}
\alias{compare_phase}
\title{Compare dose-response functions}
\usage{
compare_phase(
  lower = 60,
  upper = 215,
  mu = 140,
  phi = 25,
  sigma = 20,
  psi = 0.5,
  omega = 0.5,
  nu1 = 90,
  nu2 = 178,
  alpha = 140,
  tau = 10,
  npts = 20,
  linecol = c("steelblue", "firebrick")
)
}
\description{
Plot monophasic and biphasic dose-response functions based on a set of input parameters.
}
\author{
Phil J. Bouchet
}
