% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_rjMCMC}
\alias{compile_rjMCMC}
\title{Dose-response curves}
\usage{
compile_rjMCMC(
  rj.object,
  by.model = FALSE,
  covariate = NULL,
  covariate.values = NULL,
  species = NULL
)
}
\arguments{
\item{rj.object}{Input rjMCMC object of class \code{rjtrace}, as returned by \code{\link{trace_rjMCMC}}.}

\item{by.model}{Logical. If \code{TRUE}, the functions subsets posterior parameter estimates to produce separate dose-response curves for each model.}

\item{covariate}{Covariate name. This argument can be used to generate dose-response curves for specific contextual covariates, conditioned on the species (group) given in \code{species}.}

\item{covariate.values}{A vector of values for which dose-response curves are required. Only valid for continuous covariates.}

\item{species}{Species name.}
}
\value{
A list object of class \code{dose_response}.
}
\description{
Compute dose-response functions from a fitted rjMCMC model.
}
\examples{
library(espresso)

# Import the example data

mydat <- read_data(file = NULL) 

# Import a real dataset with the sonar and range covariates, 
# excluding sperm whales and any other species with a sample size
# smaller than two

mydat <- read_data(file = "path/to/my/data.csv", 
                  exclude.species = "Sperm whale",
                  min.N = 2) 

}
\seealso{
\code{\link{simulate_data}} \code{\link{example_brs}} \code{\link{summary.rjdata}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{rjmcmc}
