% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_rjMCMC.R
\name{configure_rjMCMC}
\alias{configure_rjMCMC}
\title{Initial configuration}
\usage{
configure_rjMCMC(
  dat,
  model.select = TRUE,
  covariate.select = FALSE,
  function.select = TRUE,
  biphasic = FALSE,
  proposal.mh = list(t.ij = 10, mu.i = 10, mu = 7, phi = 3, sigma = 1, nu = 5, tau = 5,
    alpha = 1, mu.ij = 1, psi = 1, omega = 0.5, psi.i = 0.5, k.ij = 0.5),
  proposal.rj = list(rj = 5, dd = 10, cov = 7),
  priors = list(covariates = c(0, 30), sigma = c(0, 45), phi = c(0, 45), omega = c(0,
    2), tau = c(0, 45), psi = c(0, 1)),
  p.split = 0.5,
  p.merge = 0.5,
  moves = 0,
  move.ratio = list(dd1 = 3, dd2 = 1, random = 1),
  m = 100,
  bootstrap = TRUE,
  n.rep = 1000
)
}
\arguments{
\item{dat}{Input BRS data. Must be an object of class \code{brsdata}.}

\item{model.select}{Logical. If \code{TRUE}, between-model jumps will be allowed in the rjMCMC sampler. This argument can be set to \code{FALSE} to impose a species grouping and constrain parameter estimation accordingly.}

\item{covariate.select}{Logical. Set to \code{TRUE} to allow contextual covariates to drop out of the model.}

\item{function.select}{Logical. Set to \code{TRUE} to evaluate support for dose-response functional forms (monophasic vs. biphasic).}

\item{biphasic}{Logical. Specifies which functional form to impose. Monophasic dose-response curves are returned when \code{biphasic = FALSE}. This argument is ignored when \code{function.select = TRUE}.}

\item{proposal.mh}{Named list specifying the standard deviations of the proposal distributions used in the Metropolis-Hastings sampler.}

\item{proposal.rj}{Named list specifying the standard deviations of the proposal distributions used in the reversible jump sampler. Must contain two elements: \code{mu} for response thresholds, and \code{cov} for covariates.}

\item{priors}{Named list giving the means and standard deviations for Normal priors and the lower and upper bounds for Uniform priors. Relevant parameters for the monophasic model are as follows: \code{sigma} (Uniform), \code{phi} (Uniform). Parameters for the biphasic model include: \code{alpha} (Uniform), \code{omega} (Uniform), \code{tau} (Uniform), \code{psi} (Normal). Normal priors are placed on contextual \code{covariates}.}

\item{p.split}{Probability of choosing to split a group of species when initiating a split-merge move. This parameter is constrained to be \code{1} when all species are in a single group. \code{p.split} and \code{p.merge} must sum to \code{1}.}

\item{p.merge}{Probability of choosing to merge two groups of species when initiating a split-merge move. This parameter is constrained to be \code{1} when all species are in their own groups. \code{p.split} and \code{p.merge} must sum to \code{1}.}

\item{moves}{Integer indicating which between-model moves to implement. 0 = split/merge, 1 = data driven Type I, 2 = data driven Type II, 3 = random.}

\item{move.ratio}{Relative proportion of calls to data-driven (type I and II) and independence samplers.}

\item{m}{Integer. Frequency (every \code{m} iterations) at which data-driven (type I and II) and independence samplers are triggered.}

\item{bootstrap}{Logical, defaults to TRUE. Whether to perform bootstrap clustering. As this step can be time-consuming, setting this argument to \code{FALSE} increase efficiency when reconfiguring the sampler.}

\item{n.rep}{Number of replicate bootstrap datasets used for clustering (see \code{Details}).}
}
\value{
A list object of class \code{rjconfig} identical to the input \code{brsdata} object, with an additional \code{config} element composed of:

\tabular{ll}{
\code{var} \tab Empirical estimates of φ and σ \cr
\code{prop} \tab Standard deviations of MCMC proposal distributions \cr
\code{clust} \tab Outputs from the cluster analysis \cr
\code{boot} \tab Species group assignments for each bootstrap dataset \cr
\code{mlist} \tab Unique species group assignments \cr
\code{prior} \tab Mean and standard deviation of the Normal priors placed on covariates \cr
\code{...} \tab Other elements pulled from the \code{brsdata} object \cr
}
}
\description{
Define several parameters required to configure the rjMCMC sampler.
}
\details{
The function performs three actions:
\describe{
\item{Variance}{Returns empirical estimates of between-whale (φ) and within-whale between-exposure (σ) variation, which are needed to generate starting values for the MCMC chains}
\item{Proposals}{Defines the means and standard deviations of relevant proposal distributions and priors}
\item{Clustering}{Performs a cluster analysis using \code{n.rep} bootstrap replicates of the input data, in order to parameterise between-model jumps. The analysis is run using \code{\link[mclust]{Mclust}} on both (1) species means and (2) raw observations. These respectively provide estimates of the probability distributions of (1) unique groupings, and (2) numbers of groups, where the latter ranges from \code{n = 1} when all species belong to the same group, to \code{n = n.species} when each species is treated individually.}
}
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species 
# with less than 5 observations
mydat <- read_data(file = NULL, min.N = 5) 
summary(mydat)

# Configure the rjMCMC sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 function.select = FALSE,
                                 n.rep = 100)
summary(mydat.config)
}
}
\seealso{
\code{\link{run_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{-----------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
