% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{create_groups}
\alias{create_groups}
\title{Create groupings}
\usage{
create_groups(dat.obj, species.groups = NULL, abbrev = FALSE)
}
\arguments{
\item{dat.obj}{Input dataset. Must be an object of class \code{brsdata}, as returned by \code{\link{read_data}} or \code{\link{simulate_data}}.}

\item{species.groups}{List indicating which species should be grouped. These can be given using any combination of scientific name, common name, or unique identifier, as defined in \code{\link{species_brs}}. If \code{species.groups} is given as a named list, the names of each list element will be used as group names. If \code{species.groups} is unnamed, the function will combine species codes from \code{\link{species_brs}} to generate appropriate group names.}

\item{abbrev}{Logical. If \code{TRUE}, group names will be abbreviated. This is useful for plotting and data summaries.}
}
\value{
A list object of class \code{brsdata.group}.
}
\description{
Assign species to \emph{a priori} groupings.
}
\examples{
library(espresso)

# Import the example data, excluding Risso's dolphins and minke whales
mydat <- read_data(file = NULL, exclude.species = c("gg", "ba")) 
summary(mydat)

# Group all beaked whales together
mydat.grouped <- create_groups(mydat, 
                  species.groups = list(c("Md", "Cuvier's beaked whale", 
                                        "Hyperoodon ampullatus")))

Examine the resulting grouping
summary(mydat.grouped)

}
\seealso{
\code{\link{undo_groups}} \code{\link{summary.brsdata}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
