% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_report}
\alias{create_report}
\title{Print method for objects of class \code{gvs}}
\usage{
create_report(outdir = getwd(), filename = "espresso_report", plot.height = 3)
}
\arguments{
\item{outdir}{Output directory}

\item{filename}{Output file name}

\item{plot.height}{Height of tile plots comparing species groupings}
}
\description{
Print method for objects of class \code{gvs}
}
