% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.addGenes}
\alias{.addGenes}
\title{Add genes to the input gene set.}
\usage{
.addGenes(gset, bgset, k)
}
\arguments{
\item{gset}{Gene set to be optimized.}

\item{bgset}{Background gene set.}

\item{k}{Maximum number of genes operated in each MCMC step.}
}
\value{
list of gene sets
}
\description{
This function adds genes to the gene set to be optimized by MCMC.
}
