% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.calc_accept_prob}
\alias{.calc_accept_prob}
\title{Calcurate the acceptance probability}
\usage{
.calc_accept_prob(acc0, acc, t)
}
\arguments{
\item{acc0}{accuracy of the previouse sample.}

\item{acc}{accuracy of the candidate sample.}

\item{t}{temperature parameter.}
}
\value{
probability given as numeric value
}
\description{
This function calculates the acceptance probability for MCMC sampling.
The condition equation is based on the traditional Simulated Annealing approach.
}
