% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{.color.legend}
\alias{.color.legend}
\title{Draw a color legend.}
\usage{
.color.legend(xl, yb, xr, yt, rect_col, lcex, level)
}
\arguments{
\item{xl, yb, xr, yt}{The lower left and upper right coordinates of the rectangle of colors.}

\item{rect_col}{The colors that will fill the rectangle.}

\item{lcex}{Character expansion factor for the labels.}

\item{level}{The values corresponds to the colors.}
}
\description{
This function draws a color legend.
}
