% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.computeAccuracy}
\alias{.computeAccuracy}
\title{Compute prediction accuracy of GraphSOM clustering results.}
\usage{
.computeAccuracy(obj, r, l)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{r}{Repetition number.}

\item{l}{Learning step number.}
}
\value{
accuracy
}
\description{
This function computes prediction accuracy of 
             GraphSOM clustering results.
}
