% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.computeARI}
\alias{.computeARI}
\title{Compute adjusted rand index (ARI).}
\usage{
.computeARI(obj, r)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{r}{Repetition number.}
}
\value{
adjusted rand index
}
\description{
This function computes ARI.
}
