% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.computeNeighbor}
\alias{.computeNeighbor}
\title{Compute neighborhood.}
\usage{
.computeNeighbor(obj, itr)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{itr}{Iteration number.}
}
\value{
neighborhood range
}
\description{
This function computes neighnorhood range.
}
