% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.definePriorDistr}
\alias{.definePriorDistr}
\title{Define a prior distribution for MCMC sampling}
\usage{
.definePriorDistr(tbl, fact)
}
\arguments{
\item{tbl}{summary table of a GraphSOM clustering result.}

\item{fact}{scaling factor for computing selection probability}
}
\value{
selection probabilities given as a matrix.
}
\description{
This function defines a prior distribution for MCMC sampling.
}
