% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.delGenes}
\alias{.delGenes}
\title{Delete k genes to a gene set}
\usage{
.delGenes(gset, bgset, k)
}
\arguments{
\item{gset}{gene set to be optimized.}

\item{bgset}{background gene set.}

\item{k}{the maximum number of genes operated in each MCMC step.}
}
\value{
list of gene sets
}
\description{
This function delete k genes to a gene set to be optimized by MCMC.
}
