% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.delGeneSets}
\alias{.delGeneSets}
\title{Delete gene sets which are already sampled.}
\usage{
.delGeneSets(glist1, glist2)
}
\arguments{
\item{glist1}{List of candidate gene sets.}

\item{glist2}{List of gene sets already sampled.}
}
\value{
List
}
\description{
This function deletes gene sets which are already sampled.
}
