% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.evalGraphSOM}
\alias{.evalGraphSOM}
\title{Sub-Function of evaluating GraphSOM clustering results}
\usage{
.evalGraphSOM(obj)
}
\arguments{
\item{obj}{espresso object.}
}
\value{
espresso object
}
\description{
This sub-function evaluates GraphSOM clustering results by 
             prediction accuracy and adjusted rand index (ARI).
}
