% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.generateDataFrameForRXMCMC}
\alias{.generateDataFrameForRXMCMC}
\title{Generate a data frame for storing replica exchange MCMC results.}
\usage{
.generateDataFrameForRXMCMC(n_repl, n)
}
\arguments{
\item{n_repl}{Number of replicas}

\item{n}{Number of rows}
}
\value{
Data frame
}
\description{
This function generates a data frame for storing MCMC results.
}
