% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{.getGO2Genes}
\alias{.getGO2Genes}
\title{Get gene information}
\usage{
.getGO2Genes(obj, ds)
}
\arguments{
\item{obj}{espresso object.}

\item{ds}{dataset to be accessed via biomaRt.}
}
\value{
list of GO and genes
}
\description{
This function gets gene information for each GO 
             via BioMart database.
}
