% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{.getGO2Offsprings}
\alias{.getGO2Offsprings}
\title{Get offsprings of each Gene Ontology}
\usage{
.getGO2Offsprings(terms)
}
\arguments{
\item{terms}{dataframe of GO terms.}
}
\value{
list of GOs and their offsprings.
}
\description{
This function gets Gene Ontology and 
             thier offsprings information via \code{GO.db}.
}
