% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{.getGO2Term}
\alias{.getGO2Term}
\title{Get Gene Ontology terms.}
\usage{
.getGO2Term(ds)
}
\arguments{
\item{ds}{Dataset to be accessed via biomaRt.}
}
\value{
Data frame whose the first and second columns indicate 
        GO ID and terms, respectively.
}
\description{
This function gets GO terms via BioMart database.
}
