% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.initBMU}
\alias{.initBMU}
\title{Initialize Best Matching Units (BMUs)}
\usage{
.initBMU(obj)
}
\arguments{
\item{obj}{espresso object}
}
\value{
espresso object
}
\description{
This function initializes BMUs.
}
