% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.initGraphSOM}
\alias{.initGraphSOM}
\title{Initialize an \code{espresso} object.}
\usage{
.initGraphSOM(
  obj,
  gset,
  radius,
  lsteps,
  stochastic,
  rmin,
  rmax,
  map_method,
  bmu_method,
  coef,
  nmin,
  nmax,
  swap
)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{gset}{Gene set.}

\item{radius}{Initial value of learning radius 
(default: the maximum distance of input toplology).}

\item{lsteps}{Number of learning steps.}

\item{stochastic}{If \code{stochastic = TRUE}, 
 GraphSOM clustering introduces stochasticity 
for map update (default: TRUE).}

\item{rmin}{Minimum value of scale factor for the stochastic map update.}

\item{rmax}{Maximum value of scale factor for the stochastic map update.}

\item{map_method}{Map initialization method.
The map vector is initilaized 
by input data if 'sample' is selected. 
Otherwise, it is initialized by 
uniformly distributed random values (default: 'sample').}

\item{bmu_method}{BMU selection method.
If the unit adjacent to the BMU candidate is empty,
it is set as BMU ('fill'). 
Otherwise, the most similar unit to input data 
is selected as BMU (default: 'min').}

\item{coef}{Coefficient of ARI for score computation.}

\item{nmin}{Minimum number of genes can be analyzed.}

\item{nmax}{Maximum number of genes can be analyzed.}

\item{swap}{Logical value determins whether to swap domains or not (default: TRUE)}
}
\value{
\code{espresso} object
}
\description{
This function initializes an espresso object 
             for GraphSOM clustering.
}
