% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.initMapByRandom}
\alias{.initMapByRandom}
\title{Initialize map by uniformly distributed random variables.}
\usage{
.initMapByRandom(obj)
}
\arguments{
\item{obj}{The \code{espresso} object}
}
\value{
\code{espresso} object
}
\description{
This function initializes map 
             by generating uniformly random values.
}
