% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancemap.R
\name{.modLabelPositions}
\alias{.modLabelPositions}
\title{Modify label positions.}
\usage{
.modLabelPositions(p)
}
\arguments{
\item{p}{\code{pheatmap} object.}
}
\description{
This function modifies positions of labels
}
