% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.replaceGenes}
\alias{.replaceGenes}
\title{Replace k genes}
\usage{
.replaceGenes(gset, bgset, k)
}
\arguments{
\item{gset}{gene set to be optimized.}

\item{bgset}{background gene set.}

\item{k}{the maximum number of genes operated in each MCMC step.}
}
\value{
list of gene sets
}
\description{
This function replace k genes
}
