% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.sampleGeneSet}
\alias{.sampleGeneSet}
\title{Sample a gene set from candidate gene sets.}
\usage{
.sampleGeneSet(probs)
}
\arguments{
\item{probs}{Probability distribution.}
}
\value{
Index number of selected sample given as an integer.
}
\description{
This function samples a gene set from the candidate gene sets 
             under the defined probability distribution.
}
