% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.searchBMUByMin}
\alias{.searchBMUByMin}
\title{Search for the best matching units (BMU).}
\usage{
.searchBMUByMin(obj)
}
\arguments{
\item{obj}{The \code{espresso} object}
}
\value{
BMUs given as a matrix.
}
\description{
This function searches for the BMUs for each input sample.
}
