% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.summarizeScores}
\alias{.summarizeScores}
\title{Summarize GraphSOM clustering results.}
\usage{
.summarizeScores(obj)
}
\arguments{
\item{obj}{The \code{espresso} object.}
}
\value{
\code{espresso} object
}
\description{
This function computs the mean, variance, maximum, and minimum 
             of scores, predicion accuracies and ARIs.
}
