% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.swapClusters}
\alias{.swapClusters}
\title{Swap clusters.}
\usage{
.swapClusters(obj, bmu, r, l)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{bmu}{Best matching units.}

\item{r}{Repetition number.}

\item{l}{Learning step number.}
}
\value{
\code{espresso} object
}
\description{
This function swap clusters on GraphSOM.
}
