% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.trainGraphSOM}
\alias{.trainGraphSOM}
\title{Sub-function of learning input data.}
\usage{
.trainGraphSOM(obj)
}
\arguments{
\item{obj}{The \code{espresso} object}
}
\value{
\code{espresso} object
}
\description{
This sub-function trains GraphSOM with input learning data.
}
