% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtnorm}
\alias{dtnorm}
\title{The Truncated Normal Distribution}
\usage{
dtnorm(x, location = 0, scale = 1, log = FALSE, L = -Inf, U = Inf)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{location}{Vector of means.}

\item{scale}{Vector of standard deviations.}

\item{log}{Logical. If \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{L}{Lower limit of the distribution.}

\item{U}{Upper limit of the distribution.}
}
\description{
Probability density for a truncated Normal distribution with mean equal to \code{location} and standard deviation equal to \code{scale}.
}
