% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espresso.R
\docType{class}
\name{Espresso-class}
\alias{Espresso-class}
\alias{Espresso}
\title{Class to define a GraphSOM object}
\description{
Define a GraphSOM object to be trained.
}
\section{Slots}{

\describe{
\item{\code{exprs}}{Expression matrix (row: samples, col: features).}

\item{\code{exprs_som}}{Data matrix for graphSOM computation (row: samples, col: features).}

\item{\code{rept_som}}{Repeat number for GraphSOM clustering.}

\item{\code{topology}}{Adjacent matrix representing the toplology of domains.}

\item{\code{dist}}{Distance matrix corresponding to \code{topology}.}

\item{\code{map}}{Values of the map's vector given by a matrix.}

\item{\code{map_method}}{Method of map initialization.}

\item{\code{radius}}{Initial value of learning radius.}

\item{\code{lsteps}}{Number of learning steps.}

\item{\code{rept}}{Number of repeats for GraphSOM clustering.}

\item{\code{bmu}}{Best matchiing units (BMU) through the learnig steps.}

\item{\code{bmu_method}}{BMU selection method.}

\item{\code{stochastic}}{If \code{stochastic = TRUE}, the stochastic strategy is introduced to map update.}

\item{\code{rmin}}{Minimum value of scale factor for the stochastic map update. 
If \code{stochastic = FALSE}, \code{rmin} is ignored.}

\item{\code{rmax}}{Maximum value of scale factor for the stochastic map update. 
If \code{stochastic = FALSE}, \code{rmax} is ignored.}

\item{\code{domain2name}}{Data frame indicating the domain IDs and their names.}

\item{\code{gset}}{Feature gene set.}

\item{\code{ssets}}{Sample sets.}

\item{\code{asgmt}}{Domain assignment.}

\item{\code{score}}{Score of GraphSOM clustering results.}

\item{\code{summary}}{Summary of GraphSOM clustering results.}

\item{\code{mcmc}}{History of replica exchange MCMC results.}

\item{\code{seed}}{Random seed.}

\item{\code{coef}}{Coefficient of ARI for score computation.}

\item{\code{nmin}}{Minimum number of genes can be analyzed.}

\item{\code{nmax}}{Maximum number of genes can be analyzed.}

\item{\code{swap}}{Logical value determins whether to swap domains or not.}

\item{\code{umap}}{UMAP result.}
}}

