\name{ESPRESSO}

\alias{ESPRESSO}

\docType{package}

\title{Package for power analysis and sample size calculation}

\description{
	Package to estimate sample-size and power by exploring simulated 
 study outcomes. It supports simulation-based power calculation 
 for stand-alone case-control studies and for case-control 
 analyses nested in cohort studies, that take account of realistic 
 assessment error.
}

\author{
Gaye, A. & Burton P.

Maintainer: Amadou Gaye <ag239@le.ac.uk> 
}

\references{
  Burton, P.R., Hansell, A.L., Fortier, I., Manolio, T.A., Khoury, 
  M.J., Little, J. & Elliott, P. 2009, \code{Size matters: just 
  how big is BIG?: Quantifying realistic sample size requirements 
  for human genome epidemiology}, International journal of 
  epidemiology, \bold{vol. 38, no. 1}, pp.263-273.
}

