% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espresso-package.R
\docType{package}
\name{espresso}
\alias{espresso}
\title{Bayesian inference for multi-species behavioural dose-response functions}
\description{
The \code{espresso} package provides methods for fitting and selecting among behavioural dose-response models using Bayesian reversible jump Markov Chain Monte Carlo (rjMCMC).
}
\details{
The package was developed to explore similarities in patterns of responsiveness across a range of cetacean species exposed to military sonar, and forms an output from the Double Mocha project led by the University of St Andrews and Duke University. For more details, see \href{https://synergy.st-andrews.ac.uk/mocha/}{https://synergy.st-andrews.ac.uk/mocha/}.
}
\author{
Phil J. Bouchet
}
