% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{getGO}
\alias{getGO}
\title{Get Gene Ontology information}
\usage{
getGO(species, gid, version = NULL)
}
\arguments{
\item{species}{species name used in biomaRt (e.g., mmusculus, hsapiens).}

\item{gid}{gene identifier (e.g., mgi_symbol, hgnc_symbol).}

\item{version}{Ensembl release version. 
If \code{versopn} is NULL, the latest version is downloaded.}
}
\value{
list of GO information
}
\description{
This function gets Gene Ontology (GO) information via 
             BioMart and Ensembl databases.
}
