% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{gibbs}
\alias{gibbs}
\title{Posterior inference for dose-response models}
\usage{
gibbs(
  dat,
  random.effects = FALSE,
  pseudo.n = 10000,
  mcmc.n = 1000,
  burnin = 1000,
  n.chains = 1,
  thin = 1,
  epsilon.upper = 30
)
}
\arguments{
\item{dat}{Input data. Must be an object of class \code{rjtrace} or \code{brsdata}.}

\item{random.effects}{Logical. When \code{TRUE}, uses a random effect model formulation.}

\item{pseudo.n}{Number of iterations for the pseudo-priors.}

\item{mcmc.n}{Number of posterior samples.}

\item{n.chains}{Number of MCMC chains.}

\item{thin}{Thinning interval.}

\item{epsilon.upper}{Upper bound on the ε parameter used in the random effect model formulation.}
}
\value{
A list object of class \code{gvs}.
}
\description{
Fits the Bayesian hierarchical dose-response model of \insertCite{Miller2014;textual}{espresso} to multiple species using \code{\link[rjags:rjags-package]{rjags}} \insertCite{Plummer2019}{espresso}, and estimates posterior model probabilities using a Gibbs Variable Selection (GVS) approach \insertCite{OHara2009}{espresso}.
}
\details{
Adapted from original code developed by Dina Sadykova as part of the \href{https://synergy.st-andrews.ac.uk/mocha/}{Mocha} project. The function can accommodate species/species groups either as a fixed or a random effect.
}
\examples{
\dontrun{
library(espresso)

# Simulate data for two species
mydat <- simulate_data(n.species = 2, 
                       n.whales = 16, 
                       max.trials = 3, 
                       covariates = list(exposed = c(0, 5), range = 0.5),
                       mu = c(101, 158), 
                       phi = 20, 
                       sigma = 20, 
                       Rc = c(210, 211), 
                       seed = 58697)
summary(mydat)
     
# Model selection by GVS                        
gvs <- gibbs(dat = mydat, 
             random.effects = FALSE, 
             include.covariates = FALSE, 
             mcmc.n = 1000, 
             burnin = 500)
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{summary.gvs}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{-------------------------------------------------}
\keyword{--------------------------------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{gvs}
