% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexa2hex}
\alias{hexa2hex}
\title{Mimick transparent colours}
\usage{
hexa2hex(input.colour, opacity, bg.colour = "white")
}
\arguments{
\item{input.colour}{Initial colour.}

\item{opacity}{Desired level of transparency (number between 0 and 1).}

\item{bg.colour}{Colour of the background. Defaults to 'white'.}
}
\description{
Function to find the HEX colour code corresponding to an input colour
with a set opacity level (i.e. emulate transparency)
}
