% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houser_dose.R
\docType{data}
\name{houser_dose}
\alias{houser_dose}
\title{Dose-response curve for California sea lions and bottlenose dolphins}
\usage{
data(houser_dose)
}
\description{
Derived from \insertCite{Houser2013a;textual}{espresso} and \insertCite{Houser2013b;textual}{espresso}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
