% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{initGraphSOM}
\alias{initGraphSOM}
\title{Initialize settings for GraohSOM clustering.}
\usage{
initGraphSOM(
  obj,
  nsamples = NULL,
  rept = 1,
  radius = NULL,
  lsteps = 100,
  stochastic = TRUE,
  rmin = 0.5,
  rmax = 1,
  nmin = 1,
  nmax = NULL,
  map_method = "sample",
  bmu_method = "min",
  coef = 1,
  swap = TRUE,
  seed = NULL
)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{nsamples}{Number of samples of each domain.}

\item{rept}{Number of replications (default: 1).}

\item{radius}{Initial value of learning radius.}

\item{lsteps}{Number of learning steps (default: 100).}

\item{stochastic}{GraphSOM clustering introduces stochasticity 
for map update if \code{TRUE} (default: TRUE).}

\item{rmin}{Minimum value of scale factor. 
for the stochastic map update (default: 0.5).}

\item{rmax}{Maximum value of scale factor 
for the stochastic map update (default: 1.0).}

\item{nmin}{Minimum number of genes can be analyzed.}

\item{nmax}{Maximum number of genes can be analyzed.}

\item{map_method}{Map initialization method. 
The map vector is initilaized 
by input learning data if 'sample' is selected.
Otherwise, it is initialized uniformly random 
(default: 'sample').}

\item{bmu_method}{If the unit adjacent to the BMU candidate is empty,
it is set as BMU ('fill'). 
Otherwise, the most similar unit to input data is 
selected as BMU (default: 'min').}

\item{coef}{Coefficient of ARI for score computation.}

\item{swap}{Logical value determins whether to swap domains or not (default: TRUE)}

\item{seed}{Random seed.}
}
\value{
espresso object
}
\description{
This function initializes the setting for GraphSOM clustering
}
