% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobson_dose.R
\docType{data}
\name{jacobson_dose}
\alias{jacobson_dose}
\title{Dose-response curve for Blainville's beaked whale}
\usage{
data(jacobson_dose)
}
\description{
Derived from Jacobson et al. (2019). “Quantifying the response of Blainville’s beaked whales to Naval sonar exercises in Hawaii."
}
\keyword{datasets}
