% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood}
\usage{
likelihood(
  rj.obj,
  iter,
  model,
  values = list(),
  include.cov = NULL,
  param.name = NULL,
  RJ = FALSE,
  lprod = TRUE
)
}
\arguments{
\item{rj.obj}{rjMCMC object.}

\item{iter}{Current iteration.}

\item{model}{Model ID number.}

\item{values}{List of proposed values, if available.}

\item{include.cov}{Boolean vector indicating which contextual covariates are included in the current iteration of the rjMCMC sampler.}

\item{RJ}{Logical. If \code{TRUE}, returns the likelihoods associated with between-model jumps. If \code{FALSE}, returns the likelihoods associated with the update step of the Metropolis sampler.}

\item{lprod}{If \code{TRUE}, returns the product of individual likelihoods.}

\item{param}{Paramter name. Ignored if \code{values} is specified.}
}
\description{
Calculate log-likelihoods required by the rjMCMC sampler.
}
