% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretti_dose.R
\docType{data}
\name{moretti_dose}
\alias{moretti_dose}
\title{Dose-response curve for Blainville's beaked whale}
\usage{
data(moretti_dose)
}
\description{
Derived from \insertCite{Moretti2014;textual}{espresso}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
