% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normal_prior}
\alias{normal_prior}
\title{Normal prior on covariates}
\usage{
normal_prior(rj.obj, param.name, param = NULL)
}
\arguments{
\item{rj.obj}{rjMCMC object.}

\item{param.name}{Parameter name}

\item{param}{Parameter values}

\item{priors}{Prior values}
}
\description{
Probability density for a Normal distribution.
}
