% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dose_response.R
\name{plot.dose_response}
\alias{plot.dose_response}
\title{Dose-response curves}
\usage{
\method{plot}{dose_response}(
  dr.object,
  which.model = 1,
  overlay = FALSE,
  colour.by = NULL,
  colour = NULL,
  colour.median = NULL,
  order.by = "species",
  show.p0_5 = TRUE,
  rotate.y = FALSE,
  all.quants = FALSE,
  scientific.name = TRUE,
  common.name = FALSE,
  overwrite.abbrev = TRUE,
  outline.95 = FALSE
)
}
\arguments{
\item{dr.object}{Input object of class \code{dose_response}.}

\item{which.model}{Integer indicating the ID of the model for which dose-response curves are to be plotted. The \code{\link{summary.dose_response}} function can be used to list model IDs, if necessary.}

\item{overlay}{Logical. If \code{TRUE}, plots will be faceted by species (or species group). If \code{FALSE}, all curves will be overlaid on the same plot.}

\item{colour.by}{By default, all dose-response curves follow the same colour scheme (\code{colour.by = NULL}). However, curves can also be coloured by \code{group} or by \code{species}.}

\item{colour}{Used to overwrite the \code{colour.by} argument, if desired.}

\item{colour.median}{Colour to use for the posterior median line. Overwrites the \code{colour.by} argument when specified.}

\item{order.by}{How should plots be ordered? When \code{order.by = "species"}, plots are ordered alphabetically. \code{order.by = "response"}, plots are ordered by response thresholds (from most to least sensitive species, as determined by posterior medians for μ. Only relevant when \code{overlay = FALSE}.}

\item{show.p0_5}{Logical. If \code{TRUE}, posterior medians are added to facet labels. Only relevant when \code{overlay = FALSE}.}

\item{rotate.y}{Logical. If \code{TRUE}, y-axis labels are rotated clockwise by 90 degrees.}

\item{all.quants}{}

\item{scientific.name}{Logical. If \code{TRUE}, use species' scientific names in plot titles.}

\item{common.name}{Logical. If \code{TRUE}, use species' common names in plot titles.}

\item{overwrite.abbrev}{Logical. If \code{TRUE}, overwrites abbreviated names for species grouped a priori using \code{\link{cerate_groups}}.}

\item{outline.95}{Logical. If \code{TRUE}, outlines the 95\% posterior credible interval bounds.}
}
\description{
Plot method for \code{dose_response} objects produced by \code{\link{compile_rjMCMC}}.
}
\examples{
library(espresso)

# Import the example data

mydat <- read_data(file = NULL) 

# Import a real dataset with the sonar and range covariates, 
# excluding sperm whales and any other species with a sample size
# smaller than two

mydat <- read_data(file = "path/to/my/data.csv", 
                  exclude.species = "Sperm whale",
                  min.N = 2) 

}
\seealso{
\code{\link{simulate_data}} \code{\link{example_brs}} \code{\link{summary.rjdata}}
}
\author{
Phil J. Bouchet
}
\keyword{brs}
\keyword{rjmcmc}
