% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gvs.R
\name{plot.gvs}
\alias{plot.gvs}
\title{Diagnostic plots}
\usage{
\method{plot}{gvs}(gvs.dat, param.name = NULL, autocorr = FALSE, individual = FALSE)
}
\arguments{
\item{gvs.dat}{Gibbs Variable Selection object, as returned by \code{\link{gibbs}}.}

\item{param.name}{Parameter name(s). Defaults to \code{NULL}, which returns plots for all parameters in the model.}

\item{autocorr}{Logical. Whether to output chain autocorrelation plots.}

\item{individual}{Logical. If \code{TRUE}, separate density lines will be plotted for each chain. If \code{FALSE}, one density line will be plotted for all chains.}
}
\description{
Generate trace, density, and autocorrelation plots from a \code{gvs} object.
}
\details{
Adapted from Casey Youngflesh's function \code{\link[MCMCvis]{MCMCtrace}}.
}
\seealso{
\code{\link{run_rjMCMC}} \code{\link{trace_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{brs}
\keyword{dose-response}
\keyword{gvs}
