% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rjtrace.R
\name{plot.rjtrace}
\alias{plot.rjtrace}
\title{Diagnostic plots}
\usage{
\method{plot}{rjtrace}(
  rj.obj,
  param.name = "all",
  covariates.incl = TRUE,
  autocorr = FALSE,
  individual = TRUE
)
}
\arguments{
\item{rj.obj}{rjMCMC trace object of class \code{rjtrace}.}

\item{param.name}{Parameter name(s). Defaults to \code{all}, which returns plots for all parameters in the model.}

\item{covariates.incl}{Logical. If \code{TRUE}, the trace is filtered to only retain posterior estimates obtained when the contextual covariates were included in the model. Only relevant when \code{covariate.select = TRUE} in \code{\link{configure_rjMCMC}}.}

\item{autocorr}{Logical. Whether to output chain autocorrelation plots.}

\item{individual}{Logical. If \code{TRUE}, separate density lines will be plotted for each chain. If \code{FALSE}, one density line will be plotted for all chains.}
}
\description{
Generate trace, density, and autocorrelation plots from an \code{rjtrace} object.
}
\details{
Adapted from Casey Youngflesh's \code{\link[MCMCvis]{MCMCtrace}}.
}
\examples{
library(espresso)

# Import the example data

mydat <- read_data(file = NULL) 

# Import a real dataset with the sonar and range covariates, 
# excluding sperm whales and any other species with a sample size
# smaller than two

mydat <- read_data(file = "path/to/my/data.csv", 
                  exclude.species = "Sperm whale",
                  min.N = 2) 

}
\seealso{
\code{\link{run_rjMCMC}} \code{\link{trace_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{rjmcmc}
