% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase.R
\name{plot_phase}
\alias{plot_phase}
\title{Compare dose-response functions}
\usage{
plot_phase(
  rjdat = NULL,
  plot.mono = TRUE,
  plot.bi = TRUE,
  lower = 60,
  upper = 215,
  mu = 140,
  phi = 25,
  sigma = 20,
  psi = 0.5,
  omega = 0.5,
  nu1 = 90,
  nu2 = 178,
  alpha = 140,
  tau1 = 10,
  tau2 = 10,
  npts = 20,
  linecol = c("steelblue", "firebrick")
)
}
\description{
Plot monophasic and biphasic dose-response functions based on a set of input parameters.
}
\author{
Phil J. Bouchet
}
