% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{plotConvCurve}
\alias{plotConvCurve}
\title{Plot convergence curve of GraphSOM clustering.}
\usage{
plotConvCurve(obj, rept = NULL, col = "deepskyblue4", type = "l", lwd = 1.5)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{rept}{Repetition number.}

\item{col}{Line color.}

\item{type}{Line type.}

\item{lwd}{Line width.}
}
\description{
This function outputs the convergence curve of 
             the GraphSOM clustering. The x-axis and y-axis indicate 
             the learning-steps and the number of cells 
             that migrated between clusters.
}
