% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Plot MCMC history.}
\usage{
plotMCMC(obj)
}
\arguments{
\item{obj}{The \code{espresso} object.}
}
\description{
This function plots the sampling history and replica exchange history.
}
