% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dose_response}
\alias{print.dose_response}
\title{Print method for objects of class \code{dose_response}}
\usage{
\method{print}{dose_response}(dr.object)
}
\arguments{
\item{dr.object}{Input object.}
}
\description{
Print method for objects of class \code{dose_response}
}
