% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.gvs}
\alias{print.gvs}
\title{Print method for objects of class \code{gvs}}
\usage{
\method{print}{gvs}(gvs.dat)
}
\arguments{
\item{gvs.dat}{Input object.}
}
\description{
Print method for objects of class \code{gvs}
}
