% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.rjtrace}
\alias{print.rjtrace}
\title{Print method for objects of class \code{rjtrace}}
\usage{
\method{print}{rjtrace}(rj.obj)
}
\arguments{
\item{rj.obj}{Input object.}
}
\description{
Print method for objects of class \code{rjtrace}
}
