% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_covariates}
\alias{prob_covariates}
\title{Posterior inclusion probabilities for contextual covariates}
\usage{
prob_covariates(input.obj, do.combine)
}
\description{
Posterior inclusion probabilities for contextual covariates
}
