% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propdens_mh}
\alias{propdens_mh}
\title{Proposal densities for between-model jumps}
\usage{
propdens_mh(rj.obj, param.name, dest, orig)
}
\arguments{
\item{param.name}{Parameter name}

\item{dest}{Proposed value(s)}

\item{orig}{Current value(s)}
}
\description{
Proposal densities for between-model jumps
}
