% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proposal_ff}
\alias{proposal_ff}
\title{Proposal for jumps between functional forms}
\usage{
proposal_ff(rj.obj, from.phase, prop.scale = list(alpha = 1, omega = 1))
}
\arguments{
\item{rj.obj}{Input list.}

\item{from.phase}{Jump from monophasic to biphasic (1) or from biphasic to monophasic (2).}

\item{scale.adj}{SD used in calls to rtnorm for the following parameters: alpha, nu, sigma, phi, mu}
}
\description{
Generate proposed values for relevant model parameters
}
