% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proposal_t.ij}
\alias{proposal_t.ij}
\title{Proposal for within-model jumps}
\usage{
proposal_t.ij()
}
\arguments{
\item{rj.obj}{RJ object.}

\item{param.name}{Parameter name.}

\item{seed}{Random seed. Only used for testing purposes.}
}
\description{
Generate proposed values for relevant model parameters
}
