% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_finder}
\alias{range_finder}
\title{Range finder}
\usage{
range_finder(rge, SL, target.L)
}
\arguments{
\item{rge}{Range in km.}

\item{SL}{Level of the noise source.}

\item{target.L}{Target noise level.}
}
\description{
Range finder
}
