% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Import dose-response data}
\usage{
read_data(
  file = NULL,
  include.species = NULL,
  exclude.species = NULL,
  min.N = NULL,
  covariates = NULL,
  sonar.groups = list(MFAS = c("MFAS", "MFA", "REAL MFA", "MFAS_DS", "MFA HELO",
    "REAL MFA HELO", "SOCAL_d", "REAL MFAS", "MF ALARM"), LFAS = NULL),
  exclude.sonar = c("PRN", "CAS", "ALARM", "HF ALARM", "HFAS", "HF ALARM CAS",
    "MFAS CAS"),
  range.dB = c(60, 215),
  range.var = c(0, 45),
  obs.sd = 2.5,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to the input CSV file. If set to \code{NULL}, imports the \code{\link{example_brs}} data.}

\item{include.species}{Character vector specifying which species should be retained. These can be selected by any combination of scientific name, common name, or unique identifier, as listed in \code{\link{species_brs}}. All species are included when this argument is set to the default of \code{NULL}.}

\item{exclude.species}{Character vector specifying which species should be discarded. These can be selected by any combination of scientific name, common name, or unique identifier, as listed in \code{\link{species_brs}}. No species are excluded when this argument is set to the default of \code{NULL}.}

\item{min.N}{Minimum number of observations per species. Species with sample sizes smaller than \code{min.N} will be removed.}

\item{covariates}{Contextual covariates. Must be a character string containing one or more of the following: \code{exposed}, \code{sonar}, \code{behaviour} or \code{range}. No covariates are considered when this argument is set to \code{NULL}.}

\item{sonar.groups}{Named list detailing which sonar signals should be grouped a priori.}

\item{exclude.sonar}{Character vector specifying which sonar signals should be excluded.}

\item{range.dB}{Bounds for the dose-response function. Must be a vector of length 2. Defaults to: (1) a lower bound of 60 dB re 1μPa, taken as a conservative lower limit of detectability given hearing sensitivity and the lowest average sea noise conditions; and (2) an upper bound of 215 dB re 1μPa at/above which all animals are expected to respond. This upper bound is consistent with the maximum source levels employed in behavioural response studies (BRSs) to date.}

\item{range.var}{Permissible range for the variance term(s) phi (between-whale variance) and sigma (between-exposure variance). Must be a vector of length 2.}

\item{obs.sd}{Measurement uncertainty (expressed as a standard deviation in received levels), in dB re 1μPa.}

\item{verbose}{Logical. Whether to print or suppress warning messages.}
}
\value{
A list object of class \code{brsdata}, with the following elements:

\tabular{ll}{
\code{dat} \tab Output dataset, after processing. \cr
\code{species} \tab Species data, including species names, groups, sample sizes etc. \cr
\code{whales} \tab Individual data, including animal IDs.  \cr
\code{trials} \tab Exposure data, including exposure IDs. \cr
\code{covariates} \tab Covariate data, including dummy coding, sonar groupings, factor levels etc. \cr
\code{obs} \tab Observations, including left- and right-censoring cutoffs.  \cr
\code{param} \tab General parameters. \cr
}
}
\description{
Read in and format data from behavioural response studies (BRSs). The function provides options for filtering the data based on species and/or sample sizes. See Details for formatting requirements.
}
\details{
The input data file must contain the following fields:

\tabular{ll}{
\code{species} \tab Species code, as listed in \code{\link{species_brs}}\cr
\code{tag_id} \tab Unique tag identifier\cr
\code{resp_SPL} \tab Sound pressure level at time of response (in dB re 1μPa) \cr
\code{max_SPL} \tab Maximum sound pressure level reached during the exposure\cr
\code{censored} \tab Integer variable indicating whether an observation is left-censored (-1), right-censored (1), or not censored (0) \cr
}

When covariates are specified, the below fields must also be included, as relevant:
\tabular{ll}{
\code{exp_order} \tab History of exposure (1 = 1st exposure, 2 = 2nd exposure, etc.)\cr
\code{exp_signal} \tab Type of sonar signal (e.g., MFAS, REAL MFA, PRN, CAS)\cr
\code{pre_feeding} \tab Behavioural mode (TRUE = feeding, FALSE = non-feeding)\cr
\code{min_range} \tab Minimum whale-source range during the exposure\cr
\code{resp_range} \tab Whale-source range at the time of response\cr
\code{inferred_resp_range} \tab Best estimate of whale-source range at the time of response\cr
\code{inferred_min_range} \tab Best estimate of minimum whale-source range during the exposure\cr
}

An example data file is included in the package. See \code{\link{example_brs}} for details.
}
\examples{
\dontrun{
library(espresso)

# Import the example data
mydat <- read_data(file = NULL) 

# Import a real dataset with the sonar and range covariates, 
# whilst excluding sperm whales and any other species with a sample size
# smaller than two observations.
mydat <- read_data(file = "path/to/my/data.csv", 
                  exclude.species = "Sperm whale",
                  min.N = 2)
}
}
\seealso{
\code{\link{simulate_data}} \code{\link{example_brs}} \code{\link{summary.brsdata}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{gvs}
\keyword{rjmcmc}
