% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtnorm}
\alias{rtnorm}
\title{The Truncated Normal Distribution}
\usage{
rtnorm(n, location, scale, L, U)
}
\arguments{
\item{n}{Number of observations.}

\item{location}{Vector of means.}

\item{scale}{Vector of standard deviations.}

\item{L}{Lower limit of the distribution.}

\item{U}{Upper limit of the distribution.}

\item{log}{Logical. If \code{TRUE}, probabilities p are given as log(p).}
}
\description{
Generate random deviates from a truncated normal distribution with mean equal to \code{location} and standard deviation equal to \code{scale}.
}
