% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rjMCMC.R
\name{run_rjMCMC}
\alias{run_rjMCMC}
\title{Reversible jump MCMC}
\usage{
run_rjMCMC(
  dat,
  n.chains = 3,
  n.burn = 1000,
  n.iter = 1000,
  do.update = FALSE,
  p.split = 0.5,
  p.merge = 0.5,
  move.ratio = list(dd1 = 3, dd2 = 1, random = 1),
  m = 100
)
}
\arguments{
\item{dat}{A configured rjMCMC object of class \code{rjconfig}, as returned by \code{\link{configure_rjMCMC}}.}

\item{n.chains}{Number of MCMC chains.}

\item{n.burn}{Number of MCMC iterations to use as burn-in.}

\item{n.iter}{Number of posterior samples.}

\item{do.update}{Logical. If \code{TRUE}, updates an existing rjMCMC object.}

\item{p.split}{Probability of choosing to split a group of species when initiating a split-merge move. This parameter is constrained to be \code{1} when all species are in a single group. \code{p.split} and \code{p.merge} must sum to \code{1}.}

\item{p.merge}{Probability of choosing to merge two groups of species when initiating a split-merge move. This parameter is constrained to be \code{1} when all species are in their own groups. \code{p.split} and \code{p.merge} must sum to \code{1}.}

\item{move.ratio}{Relative proportion of calls to data-driven (type I and II) and independence samplers.}

\item{m}{Integer. Frequency (every \code{m} iterations) at which data-driven (type I and II) and independence samplers are triggered.}
}
\value{
A list object of class \code{rjmcmc}.
}
\description{
Run the rjMCMC algorithm on dose-response data.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
}
}
\seealso{
\code{\link{configure_RJMCMC}} \code{\link{plot.rjtrace}} \code{\link{update_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{------------------------------}
\keyword{---------------------------------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
