% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{runUMAP}
\alias{runUMAP}
\title{Run UMAP.}
\usage{
runUMAP(obj, umap_param = NULL, gset = NULL, seed = 0)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{umap_param}{List of UMAP parameters (Refer to \code{uwot} package).}

\item{gset}{Gene set.}

\item{seed}{Random seed.}
}
\description{
This function runs UMAP.
}
