% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{rxmcmc}
\alias{rxmcmc}
\title{Optimize genes by replica exchange MCMC}
\usage{
rxmcmc(
  obj,
  gset = NULL,
  temp = 1,
  itr = 10,
  k = 1,
  seed = 0,
  n_cl = detectCores(),
  n_ex = 10,
  n_repl = detectCores(),
  n_ig = 3,
  fact = NULL
)
}
\arguments{
\item{obj}{espresso object}

\item{gset}{geneset}

\item{temp}{maximum temperature}

\item{itr}{the number of iterations until exchange of replicas}

\item{k}{the maximum number of genes operated in each MCMC step}

\item{seed}{the seed}

\item{n_cl}{the number of clusters for parallel computing}

\item{n_ex}{the number of exchanges}

\item{n_repl}{the number of replicas}

\item{n_ig}{the number of initial genes randomly selected for MCMC.}

\item{fact}{scaling factor for computing selection probability}
}
\value{
espresso object
}
\description{
This function performs optimization of genes included in a selected gene set.
             The optimization is done by replica exchange Markov Chain Monte Carlo (MCMC) approach.
}
