\name{sim.sesp.params}

\alias{sim.sesp.params}

\docType{data}

\title{Table of paramaters to simulate sensitivity and specificity values}

\description{
  Parameters stored in this table are used to simulate sensitivity and specificity values for genotype and enviromental measures assessment.  
}

\usage{data(sim.sesp.params)}

\format{
  A dataset of five constants
  \describe{
    \item{\code{seed.val}}{Seed value}
    \item{\code{prev.expo.geno}}{Prevalence of the genotype}
    \item{\code{prev.expo.env}}{Prevalence of the environmental exposure}
    \item{\code{R2.target}}{Squared correlation coefficient to compute sensitivity and specificty for the genetic determinant}
    \item{\code{reliability}}{Reliability to compute sensitivity and specificty for the genetic determinant}
  }
}

\examples{data(sim.sesp.params)}

\keyword{datasets}


