% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate a simulated dataset}
\usage{
simulate_data(
  n.species = 2,
  n.whales = 20,
  max.trials = 3,
  covariates = NULL,
  mu = NULL,
  phi = NULL,
  sigma = NULL,
  range.dB = c(60, 215),
  range.var = c(0, 45),
  Lc = c(100, 105),
  Rc = c(190, 195),
  obs.sd = 2.5,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{n.species}{Number of species.}

\item{n.whales}{Number of individuals. Can be specified as a vector of length \code{n.species}, allowing for uneven samples sizes across species, or as an integer multiple of \code{n.species} for equal sample sizes across species.}

\item{max.trials}{Maximum number of exposures per individual. Each animal is exposed \code{x} times, with \code{x} taken as a random draw from a Uniform distribution between 1 and \code{max.trials}.}

\item{covariates}{Contextual covariates and their associated coefficients. Must be supplied as a named list, in which the baseline levels of factor covariates are given a coefficient of zero.}

\item{mu}{Mean response threshold(s) for each species.}

\item{phi}{Between-whale variance in response thresholds.}

\item{sigma}{Within-whale between-exposure variance in response thresholds.}

\item{range.dB}{Bounds for the dose-response function. Must be a vector of length 2. Defaults to: (1) a lower bound of 60 dB re 1μPa, taken as a conservative lower limit of detectability given hearing sensitivity and the lowest average sea noise conditions; and (2) an upper bound of 215 dB re 1μPa at/above which all animals are expected to respond. This upper bound is consistent with the maximum source levels employed in behavioural response studies (BRSs) to date.}

\item{range.var}{Permissible range for the variance term(s) phi (between-whale variance) and sigma (between-exposure variance). Must be a vector of length 2.}

\item{Lc}{Left-censoring interval. Values of the minimum realised dose for each exposure are generated as random draws from a Uniform distribution within the bounds defined by `\code{Lc}.}

\item{Rc}{Right-censoring interval. Values of the maximum realised dose for each exposure are generated as random draws from a Uniform distribution within the bounds defined by `\code{Rc}.}

\item{obs.sd}{Measurement uncertainty (expressed as a standard deviation in received levels), in dB re 1μPa.}

\item{verbose}{Logical. Whether to print or suppress warning messages.}

\item{seed}{Random seed (for reproducible results).}
}
\value{
A list object of class \code{brsdata}, with the following elements:

\tabular{ll}{
\code{dat} \tab Output dataset, after processing. \cr
\code{species} \tab Species data, including species names, groups, sample sizes etc. \cr
\code{whales} \tab Individual data, including animal IDs.  \cr
\code{trials} \tab Exposure data, including exposure IDs. \cr
\code{covariates} \tab Covariate data, including dummy coding, sonar groupings, factor levels etc. \cr
\code{obs} \tab Observations, including right-censoring cutoffs.  \cr
\code{param} \tab General parameters. \cr
}
}
\description{
Simulate cetacean responses to Navy sonar exposure, as in behavioural response studies (BRSs).
}
\examples{
\dontrun{
library(espresso)

# Simulate data for two species
mydat <- simulate_data(n.species = 2, 
                       n.whales = 16, 
                       max.trials = 3, 
                       covariates = list(exposed = c(0, 5), range = 0.5),
                       mu = c(125, 158), 
                       phi = 20, 
                       sigma = 20, 
                       Rc = c(210, 211), 
                       seed = 58697)
}
}
\seealso{
\code{\link{read_data}} \code{\link{summary.brsdata}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{gvs}
\keyword{rjmcmc}
