% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_brs.R
\docType{data}
\name{species_brs}
\alias{species_brs}
\title{Cetacean species}
\format{
A tibble, with the following columns:
\tabular{ll}{
\code{Code} \tab Arbitrary species code\cr
\code{Scientific name} \tab Latin name\cr
\code{Common name} \tab Common name\cr
\code{NewCode} \tab Unique and consistent species identifier\cr
}
}
\usage{
data(species_brs)
}
\description{
List of cetacean species for which BRS data exist.
}
\keyword{datasets}
