% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.brsdata.R
\name{summary.brsdata}
\alias{summary.brsdata}
\title{Summarise dose-response data}
\usage{
\method{summary}{brsdata}(dat.obj, print.config = TRUE)
}
\arguments{
\item{dat.obj}{Input dataset. Must be an object of class \code{brsdata}, as returned by \code{\link{read_data}} or \code{\link{simulate_data}}.}

\item{print.config}{Logical. Whether to print a summary of MCMC parameters, when the data have been configured using \code{\link{configure_MCMC}}.}
}
\value{
A text-based summary of dataset properties.
}
\description{
Summary method for objects of class \code{brsdata}. Print a text-based summary of the data in the R console.
}
\examples{
library(espresso)

# Simulate data for two species
mydat <- simulate_data(n.species = 2, 
                      n.whales = 16, 
                      max.trials = 3, 
                      covariates = list(exposed = c(0, 5), range = 10)
                      mu = c(125, 158), 
                      phi = 20, 
                      sigma = 20, 
                      Rc = c(210, 211), 
                      seed = 58697)
                      
summary(mydat)
}
\seealso{
\code{\link{read_data}} \code{\link{simulate_data}}
}
\author{
Phil J. Bouchet
}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
