% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gvs.R
\name{summary.gvs}
\alias{summary.gvs}
\title{Some diagnostics for fitted dose-response models}
\usage{
\method{summary}{gvs}(
  gvs.obj,
  eff.n = TRUE,
  accept.rate = TRUE,
  convergence = TRUE,
  gelman.rubin = 1.1,
  model.ranks = TRUE,
  n.top = NULL
)
}
\arguments{
\item{gvs.obj}{Input trace object, as returned by \code{\link{gibbs}}.}

\item{eff.n}{Logical. If \code{TRUE}, returns estimates of the effective sample size for each model parameter.}

\item{accept.rate}{Logical. If \code{TRUE}, returns the acceptance rate (calculated after burn-in) for each model parameter.}

\item{convergence}{Logical. If \code{TRUE}, assesses convergence using the multivariate potential scale reduction factor (Gelman-Rubin statistic), as implemented in \code{\link[coda]{gelman.diag}}.}

\item{gelman.rubin}{Threshold for determining convergence based on the Gelman-Rubin statistic. Defaults to \code{1.1}.}

\item{model.ranks}{Logical. If \code{TRUE}, returns a summary of posterior model probabilities and associated model rankings.}

\item{n.top}{Number of top-ranking models to display when \code{model.ranks = TRUE}. Defaults to \code{NULL}, which lists all models.}
}
\value{
A detailed summary, printed to the R console.
}
\description{
Summary method for objects of class \code{gvs}, as returned by \code{\link{gibbs}}. Produces a text-based summary of: (1) effective sample sizes, (2) acceptance rates, (3) model convergence, and (4) posterior model probabilities. Note that \code{\link{gibbs}} does not implement covariate selection. As a result, posterior inclusion probabilities (PIPs) are not returned here, contrary to \code{\link{summary.rjtrace}}.
}
\examples{
\dontrun{
library(espresso)

# Simulate data for two species
mydat <- simulate_data(n.species = 2,
                       n.whales = 16,
                       max.trials = 3,
                       covariates = list(exposed = c(0, 5), range = 0.5),
                       mu = c(101, 158),
                       phi = 20,
                       sigma = 20,
                       Rc = c(210, 211),
                       seed = 58697)
summary(mydat)

# Model selection by GVS
gvs.model <- gibbs(dat = mydat,
             random.effects = FALSE,
             include.covariates = FALSE,
             mcmc.n = 1000,
             burnin = 500)

summary(gvs.model)
}
}
\seealso{
\code{\link{simulate_data}} \code{\link{example_brs}} \code{\link{summary.brsdata}}
}
\author{
Phil J. Bouchet
}
\keyword{-------------------------------------------------}
\keyword{dose-response}
\keyword{gvs}
