% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rjtrace.R
\name{summary.rjtrace}
\alias{summary.rjtrace}
\title{Some diagnostics for fitted dose-response models}
\usage{
\method{summary}{rjtrace}(
  rj.obj,
  eff.n = TRUE,
  accept.rate = TRUE,
  convergence = TRUE,
  gelman.rubin = 1.1,
  model.ranks = TRUE,
  n.top = NULL,
  covariate.prob = TRUE,
  combine.chains = FALSE
)
}
\arguments{
\item{rj.obj}{Input rjMCMC object, as returned by \code{\link{trace_rjMCMC}}.}

\item{covariate.prob}{Logical. If \code{TRUE}, returns a summary of posterior inclusion probabilities (PIPs).}

\item{combine.chains}{Logical. If \code{TRUE}, outputs are returned for each individual MCMC chain.}
}
\value{
A detailed summary, printed to the R console.
}
\description{
Summary method for objects of class \code{rjtrace}, as returned by \code{\link{trace_rjMCMC}}. Produces a text-based summary of: (1) effective sample sizes, (2) acceptance rates, (3) model convergence, (4) posterior model probabilities, and (5) posterior inclusion probabilities (PIPs) for contextual covariates (where appripriate).
}
\examples{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
summary(rj)

}
\seealso{
\code{\link{simulate_data}} \code{\link{example_brs}} \code{\link{summary.rjdata}}
}
\author{
Phil J. Bouchet
}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
