% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TL}
\alias{TL}
\title{Transmission loss}
\usage{
TL(rge, a = 0.185)
}
\arguments{
\item{rge}{Range in km.}

\item{a}{Sound absorption coefficient, in dB per km. This is frequency-dependent, and takes a value of 0.185 for a 3 kHz signal under normal sea conditions.}
}
\description{
Transmission loss
}
