% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_rjMCMC.R
\name{trace_rjMCMC}
\alias{trace_rjMCMC}
\title{Extract trace}
\usage{
trace_rjMCMC(rj.dat, burn = NULL, thin = 1)
}
\arguments{
\item{rj.dat}{MCMC object of class \code{rjmcmc}, as returned by \code{\link{run_rjMCMC}}.}

\item{burn}{Number of iterations to discard as burn-in.}

\item{thin}{Thinning interval. Retains every \code{thin} iteration(s).}
}
\value{
A list object of class \code{rjtrace} containing the trace in \code{mcmc.list} format as well as additional information needed for the derivation of dose-response curves.
}
\description{
Burn and thin the MCMC chain(s).
}
\details{
Adapted from the \code{fitR} package \url{https://github.com/sbfnk/fitR}.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)

# Burn and thin
rj.trace <- trace_rjMCMC(rj.dat = rj)
}
}
\seealso{
\code{\link{run_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
