% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_groups.R
\name{undo_groups}
\alias{undo_groups}
\title{Dissolve groupings}
\usage{
undo_groups(dat.obj)
}
\arguments{
\item{dat.obj}{Input dataset. Must be an object of class \code{brsdata.group}, as given by \code{\link{create_groups}}.}
}
\value{
A list object of class \code{brsdata}.
}
\description{
Undo species groupings generated by \code{\link{create_groups}}
}
\examples{
library(espresso)

# Import the example data
mydat <- read_data(file = NULL) 
summary(mydat)

# Group all beaked whales together
mydat.grouped <- create_groups(mydat, species.groups = list(beaked = c("Md", "Zc", "Ha")))
class(mydat.grouped)
summary(mydat.grouped)

# Undo groupings and revert back to the original data
mydat.ungrouped <- undo_groups(mydat.grouped)
class(mydat.ungrouped)
summary(mydat.ungrouped)

}
\seealso{
\code{\link{create_groups}} \code{\link{summary.brsdata}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
