% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rjMCMC.R
\name{update_rjMCMC}
\alias{update_rjMCMC}
\title{Update sampler}
\usage{
update_rjMCMC(rjdat, n.iter = 1000)
}
\arguments{
\item{rjdat}{Input rjMCMC sampler. Must be an object of class \code{rjmcmc}, as returned by \code{\link{run_rjMCMC}}.}

\item{n.iter}{Number of posterior samples.}
}
\description{
Update the Markov chain(s) associated with an existing rjMCMC sampler.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
                     
rj.update <- update_rjMCMC(rjdat = rj, n.iter = 500)
}
}
\seealso{
\code{\link{run_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{-----------------------------------------------}
\keyword{Extract}
\keyword{Make}
\keyword{Perform}
\keyword{Run}
\keyword{as}
\keyword{brs}
\keyword{checks}
\keyword{dose-response}
\keyword{existing}
\keyword{for}
\keyword{function}
\keyword{last}
\keyword{necessary}
\keyword{object}
\keyword{of}
\keyword{parameters}
\keyword{require}
\keyword{rjMCMC}
\keyword{rjmcmc}
\keyword{sampler,}
\keyword{starting}
\keyword{that}
\keyword{the}
\keyword{updated}
\keyword{updates}
\keyword{updating}
\keyword{using}
\keyword{values}
