% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{writeMCMC}
\alias{writeMCMC}
\title{Write MCMC results.}
\usage{
writeMCMC(obj, dir = NULL)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{dir}{Path to the output directory.}
}
\description{
This function outputs the results of replica exchange MCMC.
}
