% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rjMCMC.R
\name{compile_rjMCMC}
\alias{compile_rjMCMC}
\title{Dose-response curves}
\usage{
compile_rjMCMC(
  rj.object,
  by.model = FALSE,
  covariate = NULL,
  covariate.values = NULL,
  species = NULL,
  credible.intervals = seq(95, 1, by = -5)
)
}
\arguments{
\item{rj.object}{Input rjMCMC object of class \code{rjtrace}, as returned by \code{\link{trace_rjMCMC}}.}

\item{by.model}{Logical. If \code{TRUE}, the function subsets posterior parameter estimates to produce separate dose-response curves for each candidate model.}

\item{covariate}{Covariate name. This argument can be used to generate dose-response curves for specific contextual covariates, conditioned on the species (group) given by \code{species}.}

\item{covariate.values}{A vector of values for which dose-response curves are required. Only valid for continuous covariates.}

\item{species}{Species name.}

\item{credible.intervals}{Credible intervals. Must be a integer vector in \code{(0, 100]}. Defaults to 5-95\% in 5\% increments.}
}
\value{
A list object of class \code{dose_response}.
}
\description{
Compute dose-response functions from a fitted rjMCMC model.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
# Burn and thin
rj.trace <- trace_rjMCMC(rj.dat = rj)

# Get dose-response functions
doseR <- compile_rjMCMC(rj.trace)
}
}
\seealso{
\code{\link{run_rjMCMC}} \code{\link{plot.dose_response}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
