% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Generate report}
\usage{
create_report(
  outdir = getwd(),
  filename = "espresso_report",
  plot.height = 3,
  model.ranks = 1
)
}
\arguments{
\item{outdir}{Output directory. Defaults to the current working directory.}

\item{filename}{Output file name (without the file extension). Defaults to \code{espresso_report.html}.}

\item{plot.height}{Height of tile plots used for comparing species groupings.}

\item{model.ranks}{Used to define which models to get dose-response curves for when \code{by.model = TRUE}.}
}
\value{
An html file showing the outputs of \code{\link{summary.brsdata}}, \code{\link{summary.rjtrace}}, \code{\link{plot.rjtrace}}, and \code{\link{plot.dose_response}}.
}
\description{
Convenience function for creating a summary of results using R Markdown
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
                 
# Extract results
rj.post <- trace_rjMCMC(rj)

# Generate report
create_report()
}
}
\seealso{
\code{\link{summary.brsdata}} \code{\link{summary.rjtrace}} \code{\link{plot.rjtrace}} \code{\link{plot.dose_response}}
}
\author{
Phil J. Bouchet
}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
