% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.addGenes}
\alias{.addGenes}
\title{Add k genes to a gene set}
\usage{
.addGenes(gset, bgset, k)
}
\arguments{
\item{gset}{gene set to be optimized.}

\item{bgset}{background gene set.}

\item{k}{the maximum number of genes operated in each MCMC step.}
}
\value{
list of gene sets
}
\description{
This function adds k genes to a gene set to be optimized by MCMC.
}
