% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.computeAccuracy}
\alias{.computeAccuracy}
\title{Compute prediction accuracy of GraphSOM clustering results}
\usage{
.computeAccuracy(obj, r)
}
\arguments{
\item{obj}{espresso object.}

\item{r}{replication number.}
}
\value{
accuracy
}
\description{
This function computes prediction accuracy of 
             GraphSOM clustering results.
}
