% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.definePriorDistr}
\alias{.definePriorDistr}
\title{Define a probability distribution for MCMC sampling.}
\usage{
.definePriorDistr(tbl, fact)
}
\arguments{
\item{tbl}{Summary table of a GraphSOM clustering result.}

\item{fact}{Scaling factor for computing selection probability}
}
\value{
selection probabilities given as a matrix.
}
\description{
This function defines a probability distribution for MCMC sampling.
}
