% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.delGeneSets}
\alias{.delGeneSets}
\title{Delete gene sets already sampled}
\usage{
.delGeneSets(glist1, glist2)
}
\arguments{
\item{glist1}{list of candidate gene sets}

\item{glist2}{list of gene sets already sampled}
}
\value{
a list
}
\description{
This function deletes gene sets which are already sampled.
}
