% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.generateDataFrameForRXMCMC}
\alias{.generateDataFrameForRXMCMC}
\title{Generate a data frame for replica exchange MCMC results}
\usage{
.generateDataFrameForRXMCMC(n_rep, n)
}
\arguments{
\item{n_rep}{the number of replicas}

\item{n}{the number of rows}
}
\value{
a data frame
}
\description{
This function generates a data frame for storing MCMC results
}
