% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.genSampleSets}
\alias{.genSampleSets}
\title{Select samples randomly for down sampling}
\usage{
.genSampleSets(obj, nsamples, rept)
}
\arguments{
\item{obj}{espresso object.}

\item{nsamples}{the number of samples to be selected from each domain.}

\item{rept}{the number of replications.}
}
\value{
espresso object
}
\description{
This function randomly select samples 
             from each domain for the purpose of down sampling.
}
