% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{.getGO2Offsprings}
\alias{.getGO2Offsprings}
\title{Get offsprings of each Gene Ontology.}
\usage{
.getGO2Offsprings(terms)
}
\arguments{
\item{terms}{Data frame of GO terms.}
}
\value{
List of GOs and their offsprings.
}
\description{
This function gets GO and 
             thier offspring information via \code{GO.db}.
}
