% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{.getGO2Term}
\alias{.getGO2Term}
\title{Get Gene Ontology terms}
\usage{
.getGO2Term(ds)
}
\arguments{
\item{ds}{dataset to be accessed via biomaRt.}
}
\value{
data.frame whose the 1st and 2nd columns indicate 
        GO ID and terms, respectively.
}
\description{
This function gets terms of Gene Ontology 
             via BioMart database.
}
