% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.idx4ex}
\alias{.idx4ex}
\title{Make index vector for exchange of replicas}
\usage{
.idx4ex(n_repl, e)
}
\arguments{
\item{n_repl}{the number of replicas}

\item{e}{exchange time}
}
\value{
vector
}
\description{
This function makes a index vector for exchange of replicas
}
