% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.initBMU}
\alias{.initBMU}
\title{Initialize the best matching units (BMUs).}
\usage{
.initBMU(obj)
}
\arguments{
\item{obj}{The \code{espresso} object.}
}
\value{
\code{espresso} object
}
\description{
This function initializes BMUs.
}
