% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.initGraphSOM}
\alias{.initGraphSOM}
\title{Initialize an espresso object}
\usage{
.initGraphSOM(
  obj,
  gset,
  radius,
  lsteps,
  stochastic,
  rmin,
  rmax,
  map_method,
  bmu_method,
  coef,
  nmin,
  nmax
)
}
\arguments{
\item{obj}{espresso object.}

\item{gset}{vector of genes}

\item{radius}{initial value of learning radius 
(default: the maximum distance of input toplology).}

\item{lsteps}{the number of learning steps.}

\item{stochastic}{if \code{stochastic = TRUE}, 
 GraphSOM clustering introduces stochasticity 
for map update (default: TRUE).}

\item{rmin}{the minimum value of scale factor for the stochastic map update.}

\item{rmax}{the maximum value of scale factor for the stochastic map update.}

\item{map_method}{map initialization method.
The map vector is initilaized 
by input data if 'sample' is selected. 
Otherwise, it is initialized by 
uniformly distributed random values (default: 'sample').}

\item{bmu_method}{bmu selection method.
If the unit adjacent to the BMU candidate is empty,
it is set as BMU ('fill'). 
Otherwise, the most similar unit to input data 
is selected as BMU (default: 'min').}

\item{coef}{coefficient of ARI for score computation.}

\item{nmin}{the minimum number of genes can be analyzed.}

\item{nmax}{the maximum number of genes can be analyzed.}
}
\value{
espresso object
}
\description{
This function initializes an espresso object 
             for GraphSOM clustering.
}
