% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.initMapBySamples}
\alias{.initMapBySamples}
\title{Initialize map by input samples}
\usage{
.initMapBySamples(obj)
}
\arguments{
\item{obj}{espresso object}
}
\value{
espresso object
}
\description{
This function initializes map 
             by values of randomly selected samples.
}
