% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{.replaceGenes}
\alias{.replaceGenes}
\title{Replace genes of the input gene set.}
\usage{
.replaceGenes(gset, bgset, k)
}
\arguments{
\item{gset}{Gene set to be optimized.}

\item{bgset}{Background gene set.}

\item{k}{Maximum number of genes operated in each MCMC step.}
}
\value{
list of gene sets
}
\description{
This function replace genes of the input gene set.
}
