% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.searchBMUByFill}
\alias{.searchBMUByFill}
\title{Search the best matching units (BMU) by 'fill' method}
\usage{
.searchBMUByFill(obj, itr)
}
\arguments{
\item{obj}{espresso object}

\item{itr}{iteration number}
}
\value{
BMUs given as a matrix.
}
\description{
This function searches the BMUs for each input sample.
             In this function, deifferent from \code{.searchBMUByMin}, 
             if the adjacent units of the most similar unit are empty,
             one of them is selected as BMU.
}
