% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.summarizeScores}
\alias{.summarizeScores}
\title{Summarize GraphSOM clustering results}
\usage{
.summarizeScores(obj)
}
\arguments{
\item{obj}{espresso object.}
}
\value{
espresso object
}
\description{
This function computs the mean values of 
             predicion accuracies and ARIs.
}
