% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{.updateMap}
\alias{.updateMap}
\title{Update map.}
\usage{
.updateMap(obj, h, bmu)
}
\arguments{
\item{obj}{The \code{espresso} object}

\item{h}{Neighborhood range.}

\item{bmu}{Best matching units.}
}
\value{
updated map
}
\description{
This function updates map according to 
             the neighborhood function and the best matching units.
}
