% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espresso.R
\docType{class}
\name{Espresso-class}
\alias{Espresso-class}
\alias{Espresso}
\title{Class to define a GraphSOM object}
\description{
Define a GraphSOM object to be trained.
}
\section{Slots}{

\describe{
\item{\code{exprs}}{expression matrix (row: samples, col: features).}

\item{\code{exprs_som}}{data matrix for graphSOM computation (row: samples, col: features).}

\item{\code{rept_som}}{repeat number for GraphSOM.}

\item{\code{topology}}{adjacent matrix representing a domain toplology.}

\item{\code{dist}}{distance matrix corresponding to \code{topology}.}

\item{\code{map}}{map vectors given as a matrix.}

\item{\code{map_method}}{map initialization method.}

\item{\code{radius}}{initial value of learning radius (default: the maximum distance of an input toplology).}

\item{\code{lstep}}{the number of learning steps (default: #sample * 2).}

\item{\code{rept}}{the number of repeats for GraphSOM clustering.}

\item{\code{bmu}}{the best matchiing units through the learnig steps.}

\item{\code{bmu_method}}{bmu selection method.}

\item{\code{stochastic}}{if \code{stochastic = TRUE}, the stochastic strategy is introduced to map update.}

\item{\code{rmin}}{the minimum value of scale factor for the stochastic map update (default: 0.5). 
If \code{stochastic = FALSE}, \code{rmin} is ignored.}

\item{\code{rmax}}{the maximum value of scale factor for the stochastic map update (default: 1.0). 
If \code{stochastic = FALSE}, \code{rmax} is ignored.}

\item{\code{domain2name}}{corresponding table between domain IDs and their nemaes.}

\item{\code{gset}}{feature gene set.}

\item{\code{ssets}}{sample sets.}

\item{\code{asgmt}}{domain assignment.}

\item{\code{score}}{GraphSOM clustering results.}

\item{\code{summary}}{summary of GraphSOM clustering results.}

\item{\code{mcmc}}{history of MCMC results.}

\item{\code{seed}}{random seed.}

\item{\code{coef}}{coefficient of ARI for score computation.}

\item{\code{nmin}}{the minimum number of genes can be analyzed.}

\item{\code{nmax}}{the maximum number of genes can be analyzed.}
}}

