% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{graphSOM}
\alias{graphSOM}
\title{Run GraphSOM clustering}
\usage{
graphSOM(obj, gset = NULL, seed = 0, verbose = TRUE)
}
\arguments{
\item{obj}{espresso object.}

\item{gset}{vector of genes}

\item{seed}{random seed}

\item{verbose}{whether to show messages}
}
\value{
espresso object
}
\description{
This function execute GraohSOM clustering
}
