% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{graphSOM}
\alias{graphSOM}
\title{Run GraphSOM clustering.}
\usage{
graphSOM(obj, gset = NULL, seed = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{gset}{Gene set.}

\item{seed}{Random seed.}

\item{verbose}{Whether to show messages.}
}
\value{
\code{espresso} object
}
\description{
This function executes GraohSOM clustering.
}
