% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsom.R
\name{initGraphSOM}
\alias{initGraphSOM}
\title{Initialize setting for GraohSOM clustering}
\usage{
initGraphSOM(
  obj,
  nsamples = NULL,
  rept = 1,
  radius = NULL,
  lsteps = 100,
  stochastic = TRUE,
  rmin = 0.5,
  rmax = 1,
  nmin = 1,
  nmax = NULL,
  map_method = "sample",
  bmu_method = "min",
  seed = 0,
  coef = 1
)
}
\arguments{
\item{obj}{espresso object}

\item{nsamples}{the number of samples of each domain.}

\item{rept}{the number of replications (default: 1).}

\item{radius}{initial value of learning radius .
(default: the maximum distance of input toplology).}

\item{lsteps}{the number of learning steps (default: 100).}

\item{stochastic}{GraphSOM clustering introduces stochasticity 
for map update if \code{TRUE} (default: TRUE).}

\item{rmin}{the minimum value of scale factor 
for the stochastic map update (default: 0.5).}

\item{rmax}{the maximum value of scale factor 
for the stochastic map update (default: 1.0).}

\item{nmin}{the minimum number of genes can be analyzed.}

\item{nmax}{the maximum number of genes can be analyzed.}

\item{map_method}{map initialization method. 
The map vector is initilaized 
by input learning data if 'sample' is selected.
Otherwise, it is initialized uniformly random 
(default: 'sample').}

\item{bmu_method}{If the unit adjacent to the BMU candidate is empty,
it is set as BMU ('fill'). 
Otherwise, the most similar unit to input data is 
selected as BMU (default: 'min').}

\item{seed}{random seed}

\item{coef}{coefficient of ARI for score computation.}
}
\value{
espresso object
}
\description{
This function initializes the setting for GraphSOM clustering
}
