% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobson_dose.R
\docType{data}
\name{jacobson_dose}
\alias{jacobson_dose}
\title{Dose-response curve for Blainville's beaked whale}
\usage{
data(jacobson_dose)
}
\description{
Derived from \insertCite{Jacobson2019;textual}{espresso}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
