% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood}
\usage{
likelihood(
  biphasic = FALSE,
  param.name = NULL,
  rj.obj,
  values = NULL,
  included.cov = NULL,
  RJ = FALSE,
  lprod = TRUE
)
}
\arguments{
\item{biphasic}{Logical. Indicates the type of model for which likelihoods should be calculated.}

\item{param.name}{Paramter name. Ignored if \code{values} is specified.}

\item{rj.obj}{rjMCMC object.}

\item{values}{List of proposed values, if available.}

\item{included.cov}{Boolean vector indicating which contextual covariates are included in the current iteration of the rjMCMC sampler.}

\item{RJ}{Logical. If \code{TRUE}, returns the likelihoods associated with between-model jumps. If \code{FALSE}, returns the likelihoods associated with the update step of the Metropolis sampler.}

\item{lprod}{If \code{TRUE}, returns the product of individual likelihoods.}
}
\description{
Calculate the log-likelihood of a monophasic dose-response model, as required by the rjMCMC sampler.
}
