% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretti_dose.R
\docType{data}
\name{moretti_dose}
\alias{moretti_dose}
\title{Dose-response curve for Blainville's beaked whale}
\usage{
data(moretti_dose)
}
\description{
Derived from Moretti et al. (2014). “A Risk Function for Behavioral Disruption of Blainville’s Beaked Whales (Mesoplodon Densirostris) from Mid-Frequency Active Sonar.” PloS One 9 (1): e85064."
}
\keyword{datasets}
