% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dose_response.R
\name{plot.dose_response}
\alias{plot.dose_response}
\title{Dose-response curves}
\usage{
\method{plot}{dose_response}(
  dr.object,
  model.rank = 1,
  overlay = FALSE,
  do.facet = TRUE,
  colour.by = NULL,
  colour = NULL,
  colour.median = NULL,
  order.by = "response",
  show.pmed = TRUE,
  rotate.y = FALSE,
  all.credint = FALSE,
  scientific.name = FALSE,
  common.name = FALSE,
  overwrite.abbrev = TRUE,
  outline.outer = FALSE,
  do.save = FALSE,
  file.format = "pdf",
  ...
)
}
\arguments{
\item{dr.object}{Input object of class \code{dose_response}, as returned by \code{\link{compile_rjMCMC}}.}

\item{model.rank}{Integer indicating the rank of the model for which dose-response curves are to be plotted. The model with highest posterior probability is given a rank of 1.}

\item{overlay}{Logical. If \code{TRUE}, all curves will be overlaid on the same plot.}

\item{do.facet}{Logical. If \code{TRUE}, plots will be faceted by species (or species group).}

\item{colour.by}{Curves can be coloured by species group (\code{colour.by = "group"}) or by individual species (\code{colour.by = "species"}), or can use a single colour scheme (the default, \code{colour.by = NULL}).}

\item{colour}{Used to overwrite the \code{colour.by} argument, if desired.}

\item{colour.median}{Colour to use for the posterior median. Overwrites the \code{colour.by} argument.}

\item{order.by}{How should plots be arranged? Use \code{order.by = "species"} to arrange plots by species (groups) names in alphabetical order, or \code{order.by = "response"}, to arrange plots by response thresholds (from most to least sensitive species (groups), as determined by the estimated posterior medians for μ). Only relevant when \code{overlay = FALSE}.}

\item{show.pmed}{Logical. If \code{TRUE}, the values of posterior medians are added to facet labels. Only relevant when \code{overlay = FALSE}.}

\item{rotate.y}{Logical. If \code{TRUE}, y-axis labels are rotated clockwise by 90 degrees.}

\item{all.credint}{Logical. If \code{TRUE}, plot all credible intervals. If \code{FALSE}, only plot the outermost interval.}

\item{scientific.name}{Logical. If \code{TRUE}, use species' scientific names in plot titles.}

\item{common.name}{Logical. If \code{TRUE}, use species' common names in plot titles.}

\item{overwrite.abbrev}{Logical. If \code{TRUE}, overwrites abbreviated names for species grouped \emph{a priori} using \code{\link{create_groups}}.}

\item{outline.outer}{Logical. If \code{TRUE}, outline the outermost credible interval bounds.}

\item{do.save}{Logical. Whether to save plots on disk.}

\item{file.format}{Character. File format for output plots on disk. Can be one of \code{png}, \code{jpeg}, \code{tiff} or \code{pdf}. Only used when \code{do.save = TRUE}.}

\item{...}{Additional arguments as passed to \code{\link[ggplot2]{gg.save}}.}

\item{file.dir}{Path to the output directory. This defaults to the current working directory. Only used when \code{do.save = TRUE}.}
}
\description{
Plot \code{dose_response} curve objects produced by \code{\link{compile_rjMCMC}}.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 function.select = FALSE,
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
# Burn and thin
rj.trace <- trace_rjMCMC(rj.dat = rj)

# Get dose-response functions
doseR <- compile_rjMCMC(rj.trace)

# Plot dose-response curves
plot(dr.object = doseR, colour.by = "group")
}
}
\seealso{
\code{\link{compile_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{rjmcmc}
