% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rjtrace.R
\name{plot.rjtrace}
\alias{plot.rjtrace}
\title{Diagnostic plots}
\usage{
\method{plot}{rjtrace}(
  rj.obj,
  param.name = NULL,
  covariates.incl = TRUE,
  autocorr = FALSE,
  individual = FALSE
)
}
\arguments{
\item{rj.obj}{rjMCMC trace object of class \code{rjtrace}.}

\item{param.name}{Parameter name(s). Defaults to \code{all}, which returns plots for all parameters in the model.}

\item{covariates.incl}{Logical. If \code{TRUE}, the trace is filtered to only retain posterior estimates obtained when the contextual covariates were included in the model. Only relevant when \code{covariate.select = TRUE} in \code{\link{configure_rjMCMC}}.}

\item{autocorr}{Logical. Whether to output chain autocorrelation plots.}

\item{individual}{Logical. If \code{TRUE}, separate density lines will be plotted for each chain. If \code{FALSE}, one density line will be plotted for all chains.}
}
\description{
Generate trace, density, and autocorrelation plots from an \code{rjtrace} object.
}
\details{
Adapted from Casey Youngflesh's function \code{\link[MCMCvis]{MCMCtrace}}.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)

# Burn and thin
rj.trace <- trace_rjMCMC(rj.dat = rj)

# Get density and trace plots
plot(rj.trace)
}
}
\seealso{
\code{\link{run_rjMCMC}} \code{\link{trace_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{---------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
