% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancemap.R
\name{plotDistMap}
\alias{plotDistMap}
\title{Plot distance maps}
\usage{
plotDistMap(obj, type = "b")
}
\arguments{
\item{obj}{espresso object}

\item{type}{type of map ("d": domain, "s": sample, "b": both)}
}
\description{
This function plots distance matrices between domain-domain and 
             before/after GraphSOM clustering.
}
