% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot 3D UMAP}
\usage{
plotUMAP(
  obj,
  file = "umap.png",
  gset = NULL,
  movie = FALSE,
  dir = ".",
  umap_param = NULL,
  seed = 0,
  windowRect = c(0, 200, 1000, 1200),
  lit = TRUE,
  pch = 21,
  size = 0.5,
  lpch = 16,
  cex = 2,
  inset = c(0.02),
  view_mat = c(0.8662382, -0.4996285, -0.001548417, 0, 0.2102456, 0.3617015,
    0.90827781, 0, -0.4532415, -0.7871108, 0.418364525, 0, 0, 0, 0, 1),
  zoom = 0.9,
  axis = c(0, 0, 1),
  rpm = 5,
  duration = 10,
  domcol = NULL
)
}
\arguments{
\item{obj}{\code{espresso} object.}

\item{file}{path to the output directory.}

\item{gset}{vector of genes.}

\item{movie}{whether to make movie.}

\item{dir}{path to the output directory for the movie.}

\item{seed}{random seed}

\item{windowRect}{vector of four values indicating the 
left, top, right and bottom of 
the displayed window (in pixels).}

\item{lit}{specifying if lighting calculation should take place on geometry}

\item{pch}{symbols to plot}

\item{size}{size of the plotted points.}

\item{lpch}{symbols in the legend.}

\item{cex}{character expansion factor.}

\item{inset}{inset distance(s) from the margins.}

\item{view_mat}{the transformation matrix}

\item{zoom}{zoom factor.}

\item{axis}{axis of rotation.}

\item{rpm}{rotation speed in rotations per minute.}

\item{duration}{the duration of the animation.}

\item{domcol}{list indicating domains and colors.}
}
\description{
This function plots UMAP
}
