% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propdens_ff}
\alias{propdens_ff}
\title{Proposal densities for functional form jumps.}
\usage{
propdens_ff(param)
}
\arguments{
\item{param}{Parameter values.}
}
\description{
Proposal densities for functional form jumps.
}
