% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proposal_ff}
\alias{proposal_ff}
\title{Proposal for jumps between functional forms}
\usage{
proposal_ff(
  p.alpha = 10,
  p.psi.i = 0.25,
  p.omega = 1,
  p.phi = 2,
  p.sigma = 2,
  p.mu.i = 10
)
}
\arguments{
\item{rj.obj}{Input list.}

\item{from.phase}{Jump from monophasic to biphasic (1) or from biphasic to monophasic (2).}

\item{prop.scale}{SD used to generate proposed values for \code{alpha}, \code{omega} and \code{psi.i}.}
}
\description{
Generate proposed values for relevant model parameters
}
