% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proposal_mh}
\alias{proposal_mh}
\title{Proposal for within-model jumps}
\usage{
proposal_mh(rj.obj, param.name, iter, seed = NULL)
}
\arguments{
\item{param.name}{Parameter name.}

\item{seed}{Random seed. Only used for testing purposes.}
}
\description{
Generate proposed values for relevant model parameters
}
