% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proposal_rj}
\alias{proposal_rj}
\title{Proposal for between-model jumps}
\usage{
proposal_rj(jump, huelsenbeck = FALSE)
}
\arguments{
\item{jump}{Proposed model jump, as defined by \code{\link{propose_jump}}.}

\item{huelsenbeck}{Defines the type of proposal.}

\item{rj.obj}{Input list.}

\item{phase}{Monphasic (1) or biphasic (2).}
}
\description{
Generate proposed values for relevant model parameters
}
