% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proposal_rj}
\alias{proposal_rj}
\title{Proposal for between-model jumps}
\usage{
proposal_rj(rj.obj, jump, iter)
}
\arguments{
\item{jump}{Proposed model jump, as defined by \code{\link{propose_jump}}.}
}
\description{
Generate proposed values for relevant model parameters
}
