% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propose_jump}
\alias{propose_jump}
\title{Proposal for between-model jumps}
\usage{
propose_jump(rj.obj, move.type)
}
\arguments{
\item{move.type}{Type of between-model jump. (0): split/merge, (1): Bootstrap, (2): Cluster, (3): Random}
}
\description{
Propose a new model.
}
