% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rjMCMC.R
\name{run_rjMCMC}
\alias{run_rjMCMC}
\title{Reversible jump MCMC}
\usage{
run_rjMCMC(dat, n.chains = 3, n.burn = 1000, n.iter = 1000, do.update = FALSE)
}
\arguments{
\item{dat}{A configured rjMCMC object of class \code{rjconfig}, as returned by \code{\link{configure_rjMCMC}}.}

\item{n.chains}{Number of MCMC chains.}

\item{n.burn}{Number of MCMC iterations to use as burn-in.}

\item{n.iter}{Number of posterior samples.}

\item{do.update}{Logical. If \code{TRUE}, updates an existing rjMCMC object.}
}
\value{
A list object of class \code{rjmcmc}.
}
\description{
Run the rjMCMC algorithm on dose-response data.
}
\note{
A progress bar is used to monitor code execution on Mac and Linux operating systems. This feature does not currently work on Windows.
}
\examples{
\dontrun{
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 function.select = FALSE,
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
}
}
\seealso{
\code{\link{configure_RJMCMC}} \code{\link{plot.rjtrace}} \code{\link{update_rjMCMC}}
}
\author{
Phil J. Bouchet
}
\keyword{------------------------------}
\keyword{--------------------------------------}
\keyword{---------------------------------------------------}
\keyword{-----------------------------------------------------}
\keyword{brs}
\keyword{dose-response}
\keyword{rjmcmc}
