% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxmcmc.R
\name{rxmcmc}
\alias{rxmcmc}
\title{Optimize genes by replica exchange MCMC.}
\usage{
rxmcmc(
  obj,
  gset = NULL,
  temp = 1,
  itr = 10,
  k = 1,
  seed = NULL,
  n_cl = detectCores(),
  n_ex = 10,
  n_repl = detectCores(),
  n_ig = 3,
  fact = NULL,
  version = NULL
)
}
\arguments{
\item{obj}{The \code{espresso} object.}

\item{gset}{Gene set.}

\item{temp}{Maximum temperature (default: 1.0).}

\item{itr}{Number of iterations until exchange of replicas (default: 10).}

\item{k}{Maximum number of genes operated in each MCMC step (default: 1).}

\item{seed}{Random seed.}

\item{n_cl}{Number of clusters for parallel computing (default: \code{detectCores()}).}

\item{n_ex}{Number of exchanges (default: 10)}

\item{n_repl}{Number of replicas (default: \code{detectCores()}).}

\item{n_ig}{Number of initial genes randomly selected for MCMC.}

\item{fact}{Scaling factor for computing selection probability.}

\item{version}{Character that specifies on which previous version GraphSOM computation should be performed. (e.g., "0")}
}
\value{
\code{espresso} object
}
\description{
This function performs optimization of genes included in a selected gene set.
             The optimization is done by replica exchange Markov Chain Monte Carlo (MCMC) approach.
}
