% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Select important features by Boruta}
\usage{
selectFeatures(obj, maxRuns = 500)
}
\arguments{
\item{obj}{espresso object.}

\item{maxRuns}{the maximum number of importance source runs.}
}
\value{
a data.frame of the result of feature selection
}
\description{
This function get feature importance by Boruta package
}
