% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniform_prior}
\alias{uniform_prior}
\title{Uniform prior}
\usage{
uniform_prior(rj.obj, param.name, param = NULL)
}
\arguments{
\item{rj.obj}{rjMCMC object.}

\item{param.name}{Parameter name.}

\item{param}{Parameter values.}
}
\description{
Probability density for a Uniform distribution.
}
